requirements.txt
discord.py>=2.0.0
flask>=2.0.0
requests>=2.20.0
python-dotenv>=0.19.0
psutil>=5.8.0
aiohttp>=3.8.0
